proxy: split root and http contexts
